test(storage): isolate resolver coverage and drop noisy test churn
