+++ RiskDetectorApp/Tests/CloudPhoneRiskAppCoreTests/RiskReportStorageTests.swift
@@ -23,57 +23,9 @@ final class RiskReportStorageTests: XCTestCase {
 
     func testReportsDirectoryURL_ReturnsValidPath() throws {
         let url = try RiskReportStorage.reportsDirectoryURL()
+        XCTAssertTrue(url.path.contains("Application Support"))
         XCTAssertTrue(url.path.contains("CloudPhoneRiskKit"))
         XCTAssertTrue(url.path.contains("reports"))
-        XCTAssertEqual(url.lastPathComponent, "reports")
-    }
-
-    func testResolveBaseDirectoryPrefersApplicationSupport() {
-        let appSupport = URL(fileURLWithPath: "/tmp/app-support", isDirectory: true)
-        let caches = URL(fileURLWithPath: "/tmp/caches", isDirectory: true)
-        let temp = URL(fileURLWithPath: "/tmp/temp", isDirectory: true)
-
-        let resolved = RiskReportStorage.resolveBaseDirectory(
-            applicationSupportDirectories: [appSupport],
-            cachesDirectories: [caches],
-            temporaryDirectory: temp
-        )
-
-        XCTAssertEqual(
-            resolved,
-            appSupport.appendingPathComponent("CloudPhoneRiskKit", isDirectory: true)
-        )
-    }
-
-    func testResolveBaseDirectoryFallsBackToCaches() {
-        let caches = URL(fileURLWithPath: "/tmp/caches", isDirectory: true)
-        let temp = URL(fileURLWithPath: "/tmp/temp", isDirectory: true)
-
-        let resolved = RiskReportStorage.resolveBaseDirectory(
-            applicationSupportDirectories: [],
-            cachesDirectories: [caches],
-            temporaryDirectory: temp
-        )
-
-        XCTAssertEqual(
-            resolved,
-            caches.appendingPathComponent("CloudPhoneRiskKit", isDirectory: true)
-        )
-    }
-
-    func testResolveBaseDirectoryFallsBackToTemporaryDirectory() {
-        let temp = URL(fileURLWithPath: "/tmp/temp", isDirectory: true)
-
-        let resolved = RiskReportStorage.resolveBaseDirectory(
-            applicationSupportDirectories: [],
-            cachesDirectories: [],
-            temporaryDirectory: temp
-        )
-
-        XCTAssertEqual(
-            resolved,
-            temp.appendingPathComponent("CloudPhoneRiskKit", isDirectory: true)
-        )
     }
 
     // MARK: - list(in:)
